# Conflicts:
#	CHANGELOG.md
#	org.envirocar.app/build.gradle
#	org.envirocar.app/res/values/strings.xml
#	org.envirocar.app/src/org/envirocar/app/handler/TrackDAOHandler.java
#	org.envirocar.core/src/main/java/org/envirocar/core/entity/Track.java
#	org.envirocar.core/src/main/java/org/envirocar/core/entity/TrackImpl.java
#	org.envirocar.remote/src/main/java/org/envirocar/remote/RemoteModule.java
#	org.envirocar.remote/src/main/java/org/envirocar/remote/serializer/RemoteTrackListDeserializer.java
#	org.envirocar.remote/src/main/java/org/envirocar/remote/serializer/TrackSerializer.java

+++ org.envirocar.app/src/org/envirocar/app/views/statistics/GraphFragment.java
@@ -377,7 +377,7 @@ public class GraphFragment extends BaseInjectorFragment {
                         }
                         else if (choice == 1)
                         {
-                            values.set(index, values.get(index) + temp.getLength());
+                            values.set(index, values.get(index) + temp.getLength().floatValue());
                         }
 
                     } catch (Exception e) {
